requirements.txt
# ──────────────────────────────────────────────────────────────────────────────
# Music Bot — Python Requirements
# Target: Python 3.11, Ubuntu Server (Docker container)
# ──────────────────────────────────────────────────────────────────────────────

# ── Core Discord ──────────────────────────────────────────────────────────────
discord.py[voice]>=2.3.0      # Bot framework + FFmpeg voice support
PyNaCl>=1.5.0                  # Opus/NaCl voice encryption (discord.py requirement)

# ── Audio & Streaming ─────────────────────────────────────────────────────────
yt-dlp>=2024.1.0               # YouTube / audio extraction (kept current)
ytmusicapi>=1.3.0              # YouTube Music metadata

# ── Async Networking ──────────────────────────────────────────────────────────
aiohttp>=3.9.0                 # Async HTTP client (used by discord.py internally)
uvloop>=0.17.0                 # C-based event loop replacement (Linux/macOS only)
                               # Falls back to asyncio on Windows dev machines

# ── Persistence ───────────────────────────────────────────────────────────────
aiosqlite>=0.19.0              # Async SQLite wrapper for stats + guild settings

# ── Caching ───────────────────────────────────────────────────────────────────
cachetools>=5.3.0              # TTLCache used in audio/cache.py

# ── Configuration ─────────────────────────────────────────────────────────────
python-dotenv>=1.0.0           # .env file loading

# ── AI Features ───────────────────────────────────────────────────────────────
google-genai>=0.1.0            # Gemini AI responses (optional)

# ── Music Service Integration ─────────────────────────────────────────────────
spotipy>=2.22.1                # Spotify playlist support

# ── Web Dashboard ────────────────────────────────────────────────────────────
Flask>=2.3.2
Flask-SocketIO>=5.3.4
gevent>=23.9.1                 # Async worker for Flask-SocketIO (dashboard container)
gevent-websocket>=0.10.1       # WebSocket support for gevent

# ── System Monitoring ─────────────────────────────────────────────────────────
psutil>=5.9.5                  # CPU / memory metrics for dashboard

# ── HTTP Utilities ────────────────────────────────────────────────────────────
requests>=2.31.0               # Sync HTTP (used by spotipy internally)

# ── Dev & Testing ────────────────────────────────────────────────────────────
# Comment out / remove in production to keep image lean
pytest>=7.4.0
pytest-asyncio>=0.21.1
black>=23.7.0
flake8>=6.0.0